factor logic out of main.rs into lib.rs
